services/notification/pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = "Real-time notification service for Logistics ERP"
authors = [
    {name = "Logistics ERP Team"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "kafka-python>=2.0.2",
    "redis[hiredis]>=5.0.0",
    "sse-starlette>=1.8.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
